display addons, cba CfgMods
