+++ addons/main_a3/CfgMods.hpp
@@ -1,0 +1,18 @@
+class CfgMods {
+    class PREFIX {
+        author = "$STR_CBA_Author";
+        dir = "@CBA_A3";
+        name = "Community Base Addons v3.9.1";
+        picture = "x\cba\addons\main\logo_cba_ca.paa";
+        hidePicture = 1;
+        hideName = 1;
+        actionName = "Website";
+        action = "$STR_CBA_URL";
+        description = "Bugtracker: https://github.com/CBATeam/CBA_A3/issues<br />Documentation: https://github.com/CBATeam/CBA_A3/wiki";
+        logo = "logo_cba_ca.paa";
+        logoOver = "logo_cba_ca.paa";
+        tooltip = "Community Base Addons";
+        tooltipOwned = "Community Base Addons Owned";
+        overview = "What does the name Community Base Addons mean? It is a system that offers a range of features for addon-makers and mission designers. This includes a collection of community-built functions, an integrated keybinding system, and extend event-handler support for compatibility with other 3rd-party addons; and much much more.";
+    };
+};
