+++ addons/main_a3/CfgSettings.hpp
@@ -1,0 +1,11 @@
+class CfgSettings {
+    class CBA {
+        class Versioning {
+            class PREFIX {
+                class Dependencies {
+                    CBA[] = {"cba_main", {1,0,0}, "(true)"};
+                };
+            };
+        };
+    };
+};
