+++ addons/main_a3/config.cpp
@@ -14,26 +14,5 @@ class CfgPatches {
     };
 };
 
-class CfgSettings {
-    class CBA {
-        class Versioning {
-            class PREFIX {
-                class Dependencies {
-                    CBA[] = {"cba_main", {1, 0, 0}, "(true)"};
-                };
-            };
-        };
-    };
-};
-
-class CfgMods {
-    class PREFIX {
-        dir = "@CBA_A3";
-        name = "Community Base Addons (Arma III)";
-        picture = "x\cba\addons\main\logo_cba_ca.paa";
-        hidePicture = "true";
-        hideName = "true";
-        actionName = "Website";
-        action = "$STR_CBA_URL";
-    };
-};
+#include "CfgMods.hpp"
+#include "CfgSettings.hpp"
